Add 3D-printable case
This is a FreeCAD model of a 3d-printable case for the candlelight.
It is intended to be printed with ABS or PETG and to mechanically
protect the candlelight.

Boths halfs can be screwed together using 2 or 4 M3x8mm cylinder head
scrwes.

LEDs are still somewhat visible through holes in the upper part of the
case.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>